Rename patternMatrixTransforms script to bt-pattern-matrix-utils
