+++ src/bt-pattern-matrix-utils.ts
